Added script for starting the build
Modified the bblayers.conf sample file by replacing the environment
variable with ##OEROOT#. It should be compliant with yocto oe
environment and the startBuild.sh script.
Updated the DL_DIR path from the local.conf sample file to point
to the correct location.
Added startScript.sh that will initialize the build environment from
the layers directory.
This script must be started with ". ./startBuild.sh" command. Need to
update this aspect in the future.
The script will take the build directory name as a parameter, if not
will create a new one containing the current timestamp (for example:
rpi-build-yyyy-mm-dd).
The bblayers and local configuration file samples will be copied to
the conf directory for the build. The bblayers will be also modified
with the correct paths.
The last step will be to invoke bitbake for the rpi-home-image.