Updated the startBuild.sh script
If no build directory is provided as parameter to the startBuild.sh,
the script will initialize a new directory and copy the sample
configuration files and launch the bitbake command for the image.
Otherwise will only source the environment script and prompt a message
to the console to start the build manually.