Added install rules for initializing the session-bus
The dbus session-bus was not initialized since the system has no X11
support. In order to have the session bus initialized a session
configuration named session-local.conf need to be placed in
/etc/dbus-1/session.d/. After the target boots, the command:
"eval 'dbus-launch --auto-syntax'" must be run in order to start
the dbus-session.
Maybe a systemd service file will be added later to automate this job.