Various fixes bundle
In order to be able to connect to https or to send mails via SMTP
with SSL or TLS, the ca-certificates for host validation are needed.
Added CMDLINE_DEBUG = release flag in order to disable the systemd
debug messages.
Updated the piSmartHome recipe to create directory for logs of
applications, based on the latest package updates.