Fixed the resize-fs functionality
In order to have the sdcard resized to the maximum capacity at first
boot, the second partition must be mounted by kernel as rw instead
of ro. In order to have this rw was appended to cmdline.txt. This
option is configurable via the local.conf. By default the local.conf
sample has it enabled.
The second problem was that even if the script got executed the
filesystem was not resized. In order to have it resized correctly,
this operation had to be done in two steps, first step is to run fdisk
utility, remove the partition and recreate it. Then the target must
be rebooted in order to permit the kernel to load the new partition table.
The second phase starts after reboot when resize2fs utility is run.