Added tzdata and piUnitMaster to the image
In order to be able to set the timezone correctly some tzdata
packages were needed. Europe, Asia, Africa, Australia, Americas
were added. If more zones are needed some other packages need to be
included in the image.

piUnitMaster is added to the image too. This is a sample application
for the RaspberryPi based master node for RF24Mesh network.