Defined DEFAULT_TIMEZONE in local.conf
In order to set the target to the correct timezone the DEFAULT_TIMEZONE
must be specified correctly in the local.conf file. This setting
will be forwarded to the tzdata recipe and the link will be made correctly
to /etc/localtime. For example for Bucharest/Romania the correct value for
DEFAULT_TIMEZONE would be "Europe/Bucharest". In order to find what is the
correct timezone for your city you can run 'timedatectl list-timezones' on
any linux machine or consult the tzdata documentation.