Adding install rules for uiManager configuration file
For sharing the same QSettings configuration file between C++ and
QML, the file must be installed in the standard .config directory.
The organization defined in C++ will create piHome directory and
the file saved in this directory will be uiManager.conf. The qt
application will not create the path, but will create the file. The
directory must be created from the yocto recipe.